+++ utils/update-lambdas.bash
@@ -24,6 +24,22 @@ else
 fi
 STAC_SERVER_DIR="stac-server-${STAC_SERVER_TAG:1}"
 
+
+# Check if all three lambda zip files exist
+echo "Verifying lambda zip files..."
+if [ -f "lambda/api/api.zip" ] && [ -f "lambda/ingest/ingest.zip" ] && [ -f "lambda/pre-hook/pre-hook.zip" ]; then
+    echo "All lambda zip files are present"
+else
+    echo "ERROR: One or more lambda zip files are missing"
+    exit 1
+fi
+
+# remove the lambda zips. we'll test at the end to ensure they were successfully created
+rm lambda/api/api.zip
+rm lambda/ingest/ingest.zip
+rm lambda/pre-hook/pre-hook.zip
+rm modules/historical-ingest/lambda.zip
+
 echo "Downloading stac-server $STAC_SERVER_TAG to $STAC_SERVER_DIR..."
 curl -L -f --no-progress-meter -o stac-server.tgz "https://github.com/stac-utils/stac-server/archive/refs/tags/${STAC_SERVER_TAG}.tar.gz"
 tar -xzf stac-server.tgz
@@ -37,7 +53,7 @@ nvm install && nvm use
 npm install
 BUILD_PRE_HOOK=true npm run build
 cd ..
-
+ 
 echo "Copying stac-server lambdas..."
 cp "$STAC_SERVER_DIR/dist/api/api.zip" lambda/api/
 cp "$STAC_SERVER_DIR/dist/ingest/ingest.zip" lambda/ingest/
@@ -49,6 +65,15 @@ cd package
 zip -r ../../lambda.zip .
 cd ..
 zip ../lambda.zip main.py
-cd ../..
+cd ../../..
+
+# Check if all three lambda zip files exist
+echo "Verifying lambda zip files..."
+if [ -f "lambda/api/api.zip" ] && [ -f "lambda/ingest/ingest.zip" ] && [ -f "lambda/pre-hook/pre-hook.zip" ] && [ -f "modules/historical-ingest/lambda.zip" ]; then
+    echo "All lambda zip files are present"
+else
+    echo "ERROR: One or more lambda zip files are missing"
+    exit 1
+fi
 
 echo "Done!"